pyproject.toml
[project]
name = "mcp-crm"
version = "0.1.0"
description = "Servidor MCP com SQLite e busca vetorial em FAISS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp==3.1.1",
    "faiss-cpu==1.12.0",
    "numpy>=2.1,<3",
    "sentence-transformers>=5.1,<6",
]

[project.optional-dependencies]
dev = ["pytest>=8.3,<9", "ruff>=0.12,<0.13"]

[project.scripts]
mcp-crm = "mcp_crm.drivers.mcp_server:main"

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_crm"]

[tool.pytest.ini_options]
testpaths = ["testes"]
pythonpath = ["src"]
markers = [
    "smoke: exercita build e execucao via Docker com persistencia real",
]

[tool.ruff]
line-length = 100
target-version = "py312"
